HAI-2110 PR comment changes
