+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/domain/BaseHanke.kt
@@ -1,20 +1,19 @@
 package fi.hel.haitaton.hanke.domain
 
 import com.fasterxml.jackson.annotation.JsonIgnore
-import fi.hel.haitaton.hanke.AutoliikenteenKaistavaikutustenPituus
 import fi.hel.haitaton.hanke.ContactType
-import fi.hel.haitaton.hanke.Meluhaitta
-import fi.hel.haitaton.hanke.Polyhaitta
-import fi.hel.haitaton.hanke.Tarinahaitta
-import fi.hel.haitaton.hanke.Vaihe
-import fi.hel.haitaton.hanke.VaikutusAutoliikenteenKaistamaariin
 import fi.hel.haitaton.hanke.Yhteyshenkilo
+import fi.hel.haitaton.hanke.tormaystarkastelu.AutoliikenteenKaistavaikutustenPituus
+import fi.hel.haitaton.hanke.tormaystarkastelu.Meluhaitta
+import fi.hel.haitaton.hanke.tormaystarkastelu.Polyhaitta
+import fi.hel.haitaton.hanke.tormaystarkastelu.Tarinahaitta
+import fi.hel.haitaton.hanke.tormaystarkastelu.VaikutusAutoliikenteenKaistamaariin
 import java.time.ZonedDateTime
 import org.geojson.FeatureCollection
 
 interface BaseHanke : HasYhteystiedot {
     val nimi: String
-    val vaihe: Vaihe?
+    val vaihe: Hankevaihe?
     val alueet: List<Hankealue>?
     val tyomaaKatuosoite: String?
 }
