+++ services/hanke-service/src/main/kotlin/fi/hel/haitaton/hanke/domain/PublicHanke.kt
@@ -1,14 +1,12 @@
 package fi.hel.haitaton.hanke.domain
 
-import fi.hel.haitaton.hanke.AutoliikenteenKaistavaikutustenPituus
-import fi.hel.haitaton.hanke.Meluhaitta
-import fi.hel.haitaton.hanke.Polyhaitta
-import fi.hel.haitaton.hanke.Tarinahaitta
-import fi.hel.haitaton.hanke.TyomaaTyyppi
-import fi.hel.haitaton.hanke.Vaihe
-import fi.hel.haitaton.hanke.VaikutusAutoliikenteenKaistamaariin
 import fi.hel.haitaton.hanke.geometria.Geometriat
+import fi.hel.haitaton.hanke.tormaystarkastelu.AutoliikenteenKaistavaikutustenPituus
+import fi.hel.haitaton.hanke.tormaystarkastelu.Meluhaitta
+import fi.hel.haitaton.hanke.tormaystarkastelu.Polyhaitta
+import fi.hel.haitaton.hanke.tormaystarkastelu.Tarinahaitta
 import fi.hel.haitaton.hanke.tormaystarkastelu.TormaystarkasteluTulos
+import fi.hel.haitaton.hanke.tormaystarkastelu.VaikutusAutoliikenteenKaistamaariin
 import java.time.ZonedDateTime
 import org.geojson.FeatureCollection
 
@@ -55,7 +53,7 @@ data class PublicHanke(
     val kuvaus: String,
     val alkuPvm: ZonedDateTime,
     val loppuPvm: ZonedDateTime,
-    val vaihe: Vaihe,
+    val vaihe: Hankevaihe,
     val tyomaaTyyppi: MutableSet<TyomaaTyyppi>,
     val tormaystarkasteluTulos: TormaystarkasteluTulos,
     val omistajat: List<PublicHankeYhteystieto>,
